refactor: webpack configured and contentjs modularized
