+++ manifest.json
@@ -18,7 +18,7 @@
     {
       "matches": ["<all_urls>"],
       "css": ["styles.css"],
-      "js": ["content.js"]
+      "js": ["./dist/content.bundle.js"]
     }
   ]
 }
