+++ package.json
@@ -1,0 +1,5 @@
+{
+  "scripts": {
+    "build": "webpack"
+  }
+}
